% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_queue}
\alias{check_queue}
\title{Check queue input}
\usage{
check_queue(queue)
}
\arguments{
\item{queue}{Queue name, acceptable input: RANKED_SOLO_5x5, RANKED_FLEX_SR, RANKED_FLEX_TT}
}
\description{
Check queue input
}
