% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_user_id}
\alias{get_user_id}
\title{Get user's id}
\usage{
get_user_id(name, region, api, game = "any")
}
\arguments{
\item{name}{User name}

\item{region}{User's region}

\item{api}{Riot API}

\item{game}{Gamemode. Acceptable input: tft, league, any. Default to any}
}
\description{
Get user's id
}
