% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riot_key.R
\name{key}
\alias{key}
\alias{has_riot_key}
\alias{get_riot_key}
\title{Check or Get RIOT Key}
\usage{
has_riot_key()

get_riot_key()
}
\value{
logical if \code{has}, key if \code{get}
}
\description{
Check or Get RIOT Key
}
\examples{
has_riot_key()
}
