% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{league}
\alias{league}
\title{R6 Class for League of Legends end point}
\description{
R6 class to access the Teamfight Tactics API. This class is
  base on the different API path as illustrated in the Riot API website.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api}}{Riot API}

\item{\code{region}}{API region, acceptable values: BR1, EUN1, EUW1, JP1, KR, LA1, LA2, NA1, OC1, TR1, RU}

\item{\code{dry_run}}{If true all end point call will return glued URL}

\item{\code{league}}{search by league end points. See [league_league] for more info}

\item{\code{match}}{search by match end points. See [league_match] for more info}

\item{\code{summoner}}{search by summoner end points. See [league_summoner] for more info}

\item{\code{champion_mastery}}{search by champion mastery}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{league$new()}}
\item \href{#method-clone}{\code{league$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new league object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{league$new(api, region, dry_run = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api}}{Riot API}

\item{\code{region}}{Access region}

\item{\code{dry_run}}{Whether to call API}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `league` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{league$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
