% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol_challenges.R
\name{lol_get_all_challenge_percentiles}
\alias{lol_get_all_challenge_percentiles}
\title{Get All Challenge Percentiles}
\usage{
lol_get_all_challenge_percentiles(region = "na1", clean = TRUE)
}
\arguments{
\item{region}{Region routing value. Platform regions include \code{"na1"}, \code{"euw1"}, \code{"eun1"}, \code{"kr"}, \code{"br1"}, \code{"jp1"}, \code{"la1"}, \code{"la2"}, \code{"oc1"}, \code{"tr1"}, \code{"ru"}, \code{"ph2"}, \code{"sg2"}, \code{"th2"}, \code{"tw2"}, \code{"vn2"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a list of challenge percentiles
}
\description{
Get All Challenge Percentiles
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
lol_get_all_challenge_percentiles()
\dontshow{\}) # examplesIf}
}
