% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol_challenges.R
\name{lol_get_challenge_leaderboards}
\alias{lol_get_challenge_leaderboards}
\title{Get Challenge Leaderboards}
\usage{
lol_get_challenge_leaderboards(
  challenge_id,
  level,
  region = "na1",
  clean = TRUE
)
}
\arguments{
\item{challenge_id}{Challenge ID.}

\item{level}{Level. Such as \code{"MASTER"}, \code{"GRANDMASTER"}, or
\code{"CHALLENGER"}.}

\item{region}{Region routing value. Platform regions include \code{"na1"}, \code{"euw1"}, \code{"eun1"}, \code{"kr"}, \code{"br1"}, \code{"jp1"}, \code{"la1"}, \code{"la2"}, \code{"oc1"}, \code{"tr1"}, \code{"ru"}, \code{"ph2"}, \code{"sg2"}, \code{"th2"}, \code{"tw2"}, \code{"vn2"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with leaderboard entries or a list if
\code{clean = FALSE}
}
\description{
Get Challenge Leaderboards
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
lol_get_challenge_leaderboards(1, 'MASTER')
\dontshow{\}) # examplesIf}
}
