% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol_challenges.R
\name{lol_get_challenge_player_data}
\alias{lol_get_challenge_player_data}
\title{Get Challenge Player Data}
\usage{
lol_get_challenge_player_data(puuid, region = "na1", clean = TRUE)
}
\arguments{
\item{puuid}{PUUID of the player.}

\item{region}{Region routing value. Platform regions include \code{"na1"}, \code{"euw1"}, \code{"eun1"}, \code{"kr"}, \code{"br1"}, \code{"jp1"}, \code{"la1"}, \code{"la2"}, \code{"oc1"}, \code{"tr1"}, \code{"ru"}, \code{"ph2"}, \code{"sg2"}, \code{"th2"}, \code{"tw2"}, \code{"vn2"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a list with player challenge data
}
\description{
Get Challenge Player Data
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
puuid <- 'O2q8aNv6VBlFz9ABjEoOH8g1d_KO1IQHboIuog8ORd7QOH2L8nGgsYgK-TN16ucwUQsPpihH-tLtAA'
lol_get_challenge_player_data(puuid)
\dontshow{\}) # examplesIf}
}
