% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol_champion_mastery.R
\name{lol_get_champion_mastery}
\alias{lol_get_champion_mastery}
\title{Get Champion Mastery by PUUID and Champion ID}
\usage{
lol_get_champion_mastery(puuid, champion_id, region = "na1", clean = TRUE)
}
\arguments{
\item{puuid}{PUUID of the player.}

\item{champion_id}{Champion ID.}

\item{region}{Region routing value. Platform regions include \code{"na1"}, \code{"euw1"}, \code{"eun1"}, \code{"kr"}, \code{"br1"}, \code{"jp1"}, \code{"la1"}, \code{"la2"}, \code{"oc1"}, \code{"tr1"}, \code{"ru"}, \code{"ph2"}, \code{"sg2"}, \code{"th2"}, \code{"tw2"}, \code{"vn2"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with champion mastery data or a list if
\code{clean = FALSE}
}
\description{
Get Champion Mastery by PUUID and Champion ID
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
puuid <- 'O2q8aNv6VBlFz9ABjEoOH8g1d_KO1IQHboIuog8ORd7QOH2L8nGgsYgK-TN16ucwUQsPpihH-tLtAA'
lol_get_champion_mastery(puuid, 1)
\dontshow{\}) # examplesIf}
}
