% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol_champion_mastery.R
\name{lol_get_champion_mastery_score}
\alias{lol_get_champion_mastery_score}
\title{Get Champion Mastery Score by PUUID}
\usage{
lol_get_champion_mastery_score(puuid, region = "na1")
}
\arguments{
\item{puuid}{PUUID of the player.}

\item{region}{Region routing value. Platform regions include \code{"na1"}, \code{"euw1"}, \code{"eun1"}, \code{"kr"}, \code{"br1"}, \code{"jp1"}, \code{"la1"}, \code{"la2"}, \code{"oc1"}, \code{"tr1"}, \code{"ru"}, \code{"ph2"}, \code{"sg2"}, \code{"th2"}, \code{"tw2"}, \code{"vn2"}.}
}
\value{
an integer representing the total champion mastery score
}
\description{
Get Champion Mastery Score by PUUID
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
puuid <- 'O2q8aNv6VBlFz9ABjEoOH8g1d_KO1IQHboIuog8ORd7QOH2L8nGgsYgK-TN16ucwUQsPpihH-tLtAA'
lol_get_champion_mastery_score(puuid)
\dontshow{\}) # examplesIf}
}
