% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol_clash.R
\name{lol_get_clash_team}
\alias{lol_get_clash_team}
\title{Get Clash Team by ID}
\usage{
lol_get_clash_team(team_id, region = "na1", clean = TRUE)
}
\arguments{
\item{team_id}{Team ID.}

\item{region}{Region routing value. Platform regions include \code{"na1"}, \code{"euw1"}, \code{"eun1"}, \code{"kr"}, \code{"br1"}, \code{"jp1"}, \code{"la1"}, \code{"la2"}, \code{"oc1"}, \code{"tr1"}, \code{"ru"}, \code{"ph2"}, \code{"sg2"}, \code{"th2"}, \code{"tw2"}, \code{"vn2"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with team data, or \code{NULL} if not found.
Returns a list if \code{clean = FALSE}.
}
\description{
Get Clash Team by ID
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
lol_get_clash_team('PLACEHOLDER_CLASH_TEAM_ID')
\dontshow{\}) # examplesIf}
}
