% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol_league.R
\name{lol_get_grandmaster_league}
\alias{lol_get_grandmaster_league}
\title{Get Grandmaster League}
\usage{
lol_get_grandmaster_league(queue, region = "na1", clean = TRUE)
}
\arguments{
\item{queue}{Queue type. Such as \code{"RANKED_SOLO_5x5"} or \code{"RANKED_FLEX_SR"}.}

\item{region}{Region routing value. Platform regions include \code{"na1"}, \code{"euw1"}, \code{"eun1"}, \code{"kr"}, \code{"br1"}, \code{"jp1"}, \code{"la1"}, \code{"la2"}, \code{"oc1"}, \code{"tr1"}, \code{"ru"}, \code{"ph2"}, \code{"sg2"}, \code{"th2"}, \code{"tw2"}, \code{"vn2"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with league entries or a list if \code{clean = FALSE}
}
\description{
Get Grandmaster League
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
lol_get_grandmaster_league('RANKED_SOLO_5x5')
\dontshow{\}) # examplesIf}
}
