% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol_league.R
\name{lol_get_league}
\alias{lol_get_league}
\title{Get League by ID}
\usage{
lol_get_league(league_id, region = "na1", clean = TRUE)
}
\arguments{
\item{league_id}{League ID.}

\item{region}{Region routing value. Platform regions include \code{"na1"}, \code{"euw1"}, \code{"eun1"}, \code{"kr"}, \code{"br1"}, \code{"jp1"}, \code{"la1"}, \code{"la2"}, \code{"oc1"}, \code{"tr1"}, \code{"ru"}, \code{"ph2"}, \code{"sg2"}, \code{"th2"}, \code{"tw2"}, \code{"vn2"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with league entries or a list if \code{clean = FALSE}
}
\description{
Get League by ID
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
lol_get_league('7ccf5091-303b-3f5a-927b-7705b93e3564')
\dontshow{\}) # examplesIf}
}
