% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol_match.R
\name{lol_get_match}
\alias{lol_get_match}
\title{Get Match by ID}
\usage{
lol_get_match(match_id, regional = "americas", clean = TRUE)
}
\arguments{
\item{match_id}{Match ID.}

\item{regional}{Regional routing value. One of \code{"americas"}, \code{"asia"}, \code{"europe"}, \code{"sea"}, or \code{"esports"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with match data or a list if \code{clean = FALSE}
}
\description{
Get Match by ID
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
lol_get_match('NA1_5494453720')
\dontshow{\}) # examplesIf}
}
