% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol_match.R
\name{lol_get_match_ids}
\alias{lol_get_match_ids}
\title{Get Match IDs by PUUID}
\usage{
lol_get_match_ids(
  puuid,
  start = NULL,
  count = NULL,
  queue = NULL,
  type = NULL,
  start_time = NULL,
  end_time = NULL,
  regional = "americas"
)
}
\arguments{
\item{puuid}{PUUID of the player.}

\item{start}{Start index. Defaults to \code{0}.}

\item{count}{Number of match IDs to return. Defaults to \code{20}.}

\item{queue}{Queue ID to filter by.}

\item{type}{Type of match to filter by. Such as \code{"ranked"},
\code{"normal"}, \code{"tourney"}, or \code{"tutorial"}.}

\item{start_time}{Epoch timestamp in seconds to filter matches
after.}

\item{end_time}{Epoch timestamp in seconds to filter matches
before.}

\item{regional}{Regional routing value. One of \code{"americas"}, \code{"asia"}, \code{"europe"}, \code{"sea"}, or \code{"esports"}.}
}
\value{
a character vector of match IDs
}
\description{
Get Match IDs by PUUID
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
puuid <- 'O2q8aNv6VBlFz9ABjEoOH8g1d_KO1IQHboIuog8ORd7QOH2L8nGgsYgK-TN16ucwUQsPpihH-tLtAA'
lol_get_match_ids(puuid)
\dontshow{\}) # examplesIf}
}
