% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol_match.R
\name{lol_get_match_replays}
\alias{lol_get_match_replays}
\title{Get Match Replays by PUUID}
\usage{
lol_get_match_replays(puuid, regional = "americas")
}
\arguments{
\item{puuid}{PUUID of the player.}

\item{regional}{Regional routing value. One of \code{"americas"}, \code{"asia"}, \code{"europe"}, \code{"sea"}, or \code{"esports"}.}
}
\value{
a character vector of replay URLs
}
\description{
Get Match Replays by PUUID
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
puuid <- 'O2q8aNv6VBlFz9ABjEoOH8g1d_KO1IQHboIuog8ORd7QOH2L8nGgsYgK-TN16ucwUQsPpihH-tLtAA'
lol_get_match_replays(puuid)
\dontshow{\}) # examplesIf}
}
