% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol_tournament.R
\name{lol_get_tournament_code}
\alias{lol_get_tournament_code}
\title{Get Tournament Code}
\usage{
lol_get_tournament_code(tournament_code, region = "na1", clean = TRUE)
}
\arguments{
\item{tournament_code}{Tournament code.}

\item{region}{Region routing value. Platform regions include \code{"na1"}, \code{"euw1"}, \code{"eun1"}, \code{"kr"}, \code{"br1"}, \code{"jp1"}, \code{"la1"}, \code{"la2"}, \code{"oc1"}, \code{"tr1"}, \code{"ru"}, \code{"ph2"}, \code{"sg2"}, \code{"th2"}, \code{"tw2"}, \code{"vn2"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with tournament code data or a list if
\code{clean = FALSE}
}
\description{
Get Tournament Code
}
\examples{
\dontshow{if (has_riot_key() && Sys.getenv('RIOT_TOURNAMENT') != '') withAutoprint(\{ # examplesIf}
lol_get_tournament_code('tournament-code')
\dontshow{\}) # examplesIf}
}
