% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol_tournament.R
\name{lol_register_tournament}
\alias{lol_register_tournament}
\title{Register Tournament (Stub)}
\usage{
lol_register_tournament(body, region = "na1")
}
\arguments{
\item{body}{A list with \code{name} and \code{providerId} fields.}

\item{region}{Region routing value. Platform regions include \code{"na1"}, \code{"euw1"}, \code{"eun1"}, \code{"kr"}, \code{"br1"}, \code{"jp1"}, \code{"la1"}, \code{"la2"}, \code{"oc1"}, \code{"tr1"}, \code{"ru"}, \code{"ph2"}, \code{"sg2"}, \code{"th2"}, \code{"tw2"}, \code{"vn2"}.}
}
\value{
an integer tournament ID
}
\description{
Registers a tournament.
}
\examples{
\dontshow{if (has_riot_key() && Sys.getenv('RIOT_TOURNAMENT') != '') withAutoprint(\{ # examplesIf}
lol_register_tournament(
  body = list(name = 'My Tournament', providerId = 1)
)
\dontshow{\}) # examplesIf}
}
