% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lor_ranked.R
\name{lor_get_leaderboards}
\alias{lor_get_leaderboards}
\title{Get LoR Ranked Leaderboards}
\usage{
lor_get_leaderboards(regional = "americas", clean = TRUE)
}
\arguments{
\item{regional}{Regional routing value. One of \code{"americas"}, \code{"asia"}, \code{"europe"}, \code{"sea"}, or \code{"esports"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with ranked leaderboard entries or a list if
\code{clean = FALSE}
}
\description{
Get LoR Ranked Leaderboards
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
lor_get_leaderboards()
\dontshow{\}) # examplesIf}
}
