% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lor_match.R
\name{lor_get_match}
\alias{lor_get_match}
\title{Get LoR Match by ID}
\usage{
lor_get_match(match_id, regional = "americas", clean = TRUE)
}
\arguments{
\item{match_id}{Match ID.}

\item{regional}{Regional routing value. One of \code{"americas"}, \code{"asia"}, \code{"europe"}, \code{"sea"}, or \code{"esports"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with match data, or \code{NULL} if not found.
Returns a list if \code{clean = FALSE}.
}
\description{
Get LoR Match by ID
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
lor_get_match('PLACEHOLDER_LOR_MATCH_ID')
\dontshow{\}) # examplesIf}
}
