% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lor_match.R
\name{lor_get_match_ids}
\alias{lor_get_match_ids}
\title{Get LoR Match IDs by PUUID}
\usage{
lor_get_match_ids(puuid, regional = "americas")
}
\arguments{
\item{puuid}{PUUID of the player.}

\item{regional}{Regional routing value. One of \code{"americas"}, \code{"asia"}, \code{"europe"}, \code{"sea"}, or \code{"esports"}.}
}
\value{
a character vector of match IDs
}
\description{
Get LoR Match IDs by PUUID
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
puuid <- 'O2q8aNv6VBlFz9ABjEoOH8g1d_KO1IQHboIuog8ORd7QOH2L8nGgsYgK-TN16ucwUQsPpihH-tLtAA'
lor_get_match_ids(puuid)
\dontshow{\}) # examplesIf}
}
