% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{lor_get_platform_data}
\alias{lor_get_platform_data}
\title{Get LoR Platform Data (Status)}
\usage{
lor_get_platform_data(regional = "americas", clean = TRUE)
}
\arguments{
\item{regional}{Regional routing value. One of \code{"americas"}, \code{"asia"}, \code{"europe"}, \code{"sea"}, or \code{"esports"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with platform status data or a list if
\code{clean = FALSE}
}
\description{
Get LoR Platform Data (Status)
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
lor_get_platform_data()
\dontshow{\}) # examplesIf}
}
