% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riot_account.R
\name{riot_get_account_by_puuid_esports}
\alias{riot_get_account_by_puuid_esports}
\title{Get Account by PUUID (Esports)}
\usage{
riot_get_account_by_puuid_esports(puuid, regional = "esports", clean = TRUE)
}
\arguments{
\item{puuid}{PUUID of the player.}

\item{regional}{Regional routing value. One of \code{"americas"}, \code{"asia"}, \code{"europe"}, \code{"sea"}, or \code{"esports"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with account data, or \code{NULL} if not found.
Returns a list if \code{clean = FALSE}.
}
\description{
Get Account by PUUID (Esports)
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
puuid <- 'O2q8aNv6VBlFz9ABjEoOH8g1d_KO1IQHboIuog8ORd7QOH2L8nGgsYgK-TN16ucwUQsPpihH-tLtAA'
riot_get_account_by_puuid_esports(puuid)
\dontshow{\}) # examplesIf}
}
