% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riot_account.R
\name{riot_get_account_by_riot_id}
\alias{riot_get_account_by_riot_id}
\title{Get Account by Riot ID}
\usage{
riot_get_account_by_riot_id(
  game_name,
  tag_line,
  regional = "americas",
  clean = TRUE
)
}
\arguments{
\item{game_name}{Game name of the Riot account.}

\item{tag_line}{Tag line of the Riot account.}

\item{regional}{Regional routing value. One of \code{"americas"}, \code{"asia"}, \code{"europe"}, \code{"sea"}, or \code{"esports"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with account data or a list if \code{clean = FALSE}
}
\description{
Get Account by Riot ID
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
riot_get_account_by_riot_id('Doublelift', 'NA1')
\dontshow{\}) # examplesIf}
}
