% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riot_account.R
\name{riot_get_active_region}
\alias{riot_get_active_region}
\title{Get Active Region for a Player}
\usage{
riot_get_active_region(game, puuid, regional = "americas", clean = TRUE)
}
\arguments{
\item{game}{Game identifier. Such as \code{"val"} or \code{"lor"}.}

\item{puuid}{PUUID of the player.}

\item{regional}{Regional routing value. One of \code{"americas"}, \code{"asia"}, \code{"europe"}, \code{"sea"}, or \code{"esports"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with active region data, or \code{NULL} if the
request fails. Returns a list if \code{clean = FALSE}.
}
\description{
Get Active Region for a Player
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
puuid <- 'O2q8aNv6VBlFz9ABjEoOH8g1d_KO1IQHboIuog8ORd7QOH2L8nGgsYgK-TN16ucwUQsPpihH-tLtAA'
riot_get_active_region('val', puuid)
\dontshow{\}) # examplesIf}
}
