% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riotr-package.R
\docType{package}
\name{riotr-package}
\alias{riotr}
\alias{riotr-package}
\title{riotr: Interact with the 'Riot Games API'}
\description{
Download and read data from the Riot Games Developer API (\url{https://developer.riotgames.com/}). Provides functions for all primary endpoints, including League of Legends, Teamfight Tactics, Legends of Runeterra, and Valorant.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/christopherkenny/riotr}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{ctkenny@proton.me} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

}
\keyword{internal}
