% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tft.R
\name{tft}
\alias{tft}
\title{R6 Class for Teamfight Tactics end point (TFT)}
\description{
R6 Class for Teamfight Tactics end point (TFT)

R6 Class for Teamfight Tactics end point (TFT)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api}}{Riot API}

\item{\code{region}}{API region, acceptable values: BR1, EUN1, EUW1, JP1, KR, LA1, LA2, NA1, OC1, TR1, RU}

\item{\code{dry_run}}{If true all end point call will return glued URL}

\item{\code{league}}{TFT search by league end points. See [tft_league] for more info}

\item{\code{match}}{TFT search by match end points. See [tft_match] for more info}

\item{\code{summoner}}{TFT search by summoner end points. See [tft_summoner] for more info}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tft$new()}}
\item \href{#method-print}{\code{tft$print()}}
\item \href{#method-clone}{\code{tft$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new tft object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tft$new(api, region, dry_run = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api}}{Riot API}

\item{\code{region}}{Access region}

\item{\code{dry_run}}{Whether to call API}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `tft` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tft$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tft$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
