% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tft_league.R
\name{tft_get_master_league}
\alias{tft_get_master_league}
\title{Get TFT Master League}
\usage{
tft_get_master_league(queue = "RANKED_TFT", region = "na1", clean = TRUE)
}
\arguments{
\item{queue}{Queue type. Defaults to \code{"RANKED_TFT"}.}

\item{region}{Region routing value. Platform regions include \code{"na1"}, \code{"euw1"}, \code{"eun1"}, \code{"kr"}, \code{"br1"}, \code{"jp1"}, \code{"la1"}, \code{"la2"}, \code{"oc1"}, \code{"tr1"}, \code{"ru"}, \code{"ph2"}, \code{"sg2"}, \code{"th2"}, \code{"tw2"}, \code{"vn2"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with league entries or a list if \code{clean = FALSE}
}
\description{
Get TFT Master League
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
tft_get_master_league()
\dontshow{\}) # examplesIf}
}
