% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tft_match.R
\name{tft_get_match_ids}
\alias{tft_get_match_ids}
\title{Get TFT Match IDs by PUUID}
\usage{
tft_get_match_ids(
  puuid,
  start = NULL,
  count = NULL,
  start_time = NULL,
  end_time = NULL,
  regional = "americas"
)
}
\arguments{
\item{puuid}{PUUID of the player.}

\item{start}{Start index. Defaults to \code{0}.}

\item{count}{Number of match IDs to return. Defaults to \code{20}.}

\item{start_time}{Epoch timestamp in seconds to filter matches
after.}

\item{end_time}{Epoch timestamp in seconds to filter matches
before.}

\item{regional}{Regional routing value. One of \code{"americas"}, \code{"asia"}, \code{"europe"}, \code{"sea"}, or \code{"esports"}.}
}
\value{
a character vector of match IDs
}
\description{
Get TFT Match IDs by PUUID
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
puuid <- 'GmNPNjUbzbWN8YIA79V4p308eFpf21xPO2Lw8o2xJTJMPW83W54t4GfyfpwOn7-hBz5c3MHpg-Dc3w'
tft_get_match_ids(puuid)
\dontshow{\}) # examplesIf}
}
