% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tft_summoner.R
\name{tft_get_summoner_by_puuid}
\alias{tft_get_summoner_by_puuid}
\title{Get TFT Summoner by PUUID}
\usage{
tft_get_summoner_by_puuid(puuid, region = "na1", clean = TRUE)
}
\arguments{
\item{puuid}{PUUID of the player.}

\item{region}{Region routing value. Platform regions include \code{"na1"}, \code{"euw1"}, \code{"eun1"}, \code{"kr"}, \code{"br1"}, \code{"jp1"}, \code{"la1"}, \code{"la2"}, \code{"oc1"}, \code{"tr1"}, \code{"ru"}, \code{"ph2"}, \code{"sg2"}, \code{"th2"}, \code{"tw2"}, \code{"vn2"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with summoner data or a list if \code{clean = FALSE}
}
\description{
Get TFT Summoner by PUUID
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
puuid <- 'GmNPNjUbzbWN8YIA79V4p308eFpf21xPO2Lw8o2xJTJMPW83W54t4GfyfpwOn7-hBz5c3MHpg-Dc3w'
tft_get_summoner_by_puuid(puuid)
\dontshow{\}) # examplesIf}
}
