% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tft_league.R
\name{tft_get_top_rated_ladder}
\alias{tft_get_top_rated_ladder}
\title{Get TFT Top Rated Ladder}
\usage{
tft_get_top_rated_ladder(queue, region = "na1", clean = TRUE)
}
\arguments{
\item{queue}{Queue type for TFT rated ladder.}

\item{region}{Region routing value. Platform regions include \code{"na1"}, \code{"euw1"}, \code{"eun1"}, \code{"kr"}, \code{"br1"}, \code{"jp1"}, \code{"la1"}, \code{"la2"}, \code{"oc1"}, \code{"tr1"}, \code{"ru"}, \code{"ph2"}, \code{"sg2"}, \code{"th2"}, \code{"tw2"}, \code{"vn2"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with top rated ladder entries or a list if
\code{clean = FALSE}
}
\description{
Get TFT Top Rated Ladder
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
tft_get_top_rated_ladder('RANKED_TFT_TURBO')
\dontshow{\}) # examplesIf}
}
