% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tft.R
\name{tft_league}
\alias{tft_league}
\title{TFT League End Point R6 class}
\description{
Sub class for [tft] to access [TFT-LEAGUE-V1](https://developer.riotgames.com/apis#tft-league-v1)
  end points.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api}}{Riot API}

\item{\code{region}}{API region, acceptable values: BR1, EUN1, EUW1, JP1, KR, LA1, LA2, NA1, OC1, TR1, RU}

\item{\code{dry_run}}{If true all end point call will return glued URL}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tft_league$new()}}
\item \href{#method-challenger}{\code{tft_league$challenger()}}
\item \href{#method-grandmaster}{\code{tft_league$grandmaster()}}
\item \href{#method-master}{\code{tft_league$master()}}
\item \href{#method-by_summoner}{\code{tft_league$by_summoner()}}
\item \href{#method-by_tier}{\code{tft_league$by_tier()}}
\item \href{#method-by_leagueId}{\code{tft_league$by_leagueId()}}
\item \href{#method-clone}{\code{tft_league$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new tft object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tft_league$new(api, region, dry_run)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api}}{Riot API}

\item{\code{region}}{Access region}

\item{\code{dry_run}}{Whether to call API}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `tft` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-challenger"></a>}}
\if{latex}{\out{\hypertarget{method-challenger}{}}}
\subsection{Method \code{challenger()}}{
Get challenger league.
For more info see [here](https://developer.riotgames.com/apis#tft-league-v1/GET_getChallengerLeague)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tft_league$challenger(region = self$region, dry_run = self$dry_run)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{Region to query. Default to class's region. Can overwrite.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grandmaster"></a>}}
\if{latex}{\out{\hypertarget{method-grandmaster}{}}}
\subsection{Method \code{grandmaster()}}{
Get grandmaster league.
For more info see [here](https://developer.riotgames.com/apis#tft-league-v1/GET_getGrandmasterLeague)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tft_league$grandmaster(region = self$region, dry_run = self$dry_run)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{Region to query. Default to class's region. Can overwrite.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-master"></a>}}
\if{latex}{\out{\hypertarget{method-master}{}}}
\subsection{Method \code{master()}}{
Get master league.
For more info see [here](https://developer.riotgames.com/apis#tft-league-v1/GET_getMasterLeague)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tft_league$master(region = self$region, dry_run = self$dry_run)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{Region to query. Default to class's region. Can overwrite.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-by_summoner"></a>}}
\if{latex}{\out{\hypertarget{method-by_summoner}{}}}
\subsection{Method \code{by_summoner()}}{
Search client by
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tft_league$by_summoner(
  summonerName = NULL,
  region = self$region,
  dry_run = self$dry_run
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-by_tier"></a>}}
\if{latex}{\out{\hypertarget{method-by_tier}{}}}
\subsection{Method \code{by_tier()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tft_league$by_tier(
  tier = NULL,
  division = NULL,
  region = self$region,
  dry_run = self$dry_run
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-by_leagueId"></a>}}
\if{latex}{\out{\hypertarget{method-by_leagueId}{}}}
\subsection{Method \code{by_leagueId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tft_league$by_leagueId(leagueId, region = self$region, dry_run = self$dry_run)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tft_league$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
