% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val.R
\name{val_get_content}
\alias{val_get_content}
\title{Get Valorant Content}
\usage{
val_get_content(locale = NULL, regional = "na", clean = TRUE)
}
\arguments{
\item{locale}{Locale. Such as \code{"en-US"}, \code{"ja-JP"}, etc.}

\item{regional}{Regional routing value. One of \code{"americas"}, \code{"asia"}, \code{"europe"}, \code{"sea"}, or \code{"esports"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a list with Valorant content data
}
\description{
Get Valorant Content
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
val_get_content()
\dontshow{\}) # examplesIf}
}
