% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val.R
\name{val_get_leaderboard}
\alias{val_get_leaderboard}
\title{Get Valorant Ranked Leaderboard by Act}
\usage{
val_get_leaderboard(
  act_id,
  size = NULL,
  start_index = NULL,
  regional = "na",
  clean = TRUE
)
}
\arguments{
\item{act_id}{Act ID.}

\item{size}{Number of entries to return.}

\item{start_index}{Start index for pagination.}

\item{regional}{Regional routing value. One of \code{"americas"}, \code{"asia"}, \code{"europe"}, \code{"sea"}, or \code{"esports"}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with ranked leaderboard entries or a list if
\code{clean = FALSE}
}
\description{
Get Valorant Ranked Leaderboard by Act
}
\examples{
\dontshow{if (has_riot_key()) withAutoprint(\{ # examplesIf}
val_get_leaderboard('3ea2b318-423b-cf86-25da-7cbb0eefbe2d')
\dontshow{\}) # examplesIf}
}
